M2-4969: refactor app version warning banner
Now it uses the new Banner base component. Also add it to the list of
banner components available to enable in the banners state schema, and
renamed BannerSchema to BannersSchema.
