+++ src/resources/app-fr.json
@@ -1246,7 +1246,7 @@
   "fromToHint": "De {{min}} à {{max}}",
   "selectValidInterval": "Sélectionnez un intervalle valide",
   "selectValueWithinInterval": "Sélectionnez une valeur dans un intervalle",
-  "bannerText": "<0>Vous utilisez la nouvelle version de MindLogger!</0><1>Les utilisateurs finaux doivent mettre à jour vers la nouvelle application.</1><2>Prenez ces mesures dès maintenant pour vous assurer que les données de réponse des participants ne sont pas perdues.</2>",
+  "versionWarningBanner": "<0>Vous utilisez la nouvelle version de MindLogger!</0> <1>Les utilisateurs finaux doivent mettre à jour vers la nouvelle application.</1> <2>Prenez ces mesures dès maintenant pour vous assurer que les données de réponse des participants ne sont pas perdues.</2>",
   "youNeedToAuthorizeHint": "Pour créer une nouvelle applet ou la fusionner avec d'autres, vous devez autoriser.",
   "errorFallback": {
     "somethingWentWrong": "Quelque chose s'est mal passé.",
