+++ src/shared/components/Banners/VersionWarningBanner/VersionWarningBanner.const.ts
