+++ src/shared/components/Banners/VersionWarningBanner/index.ts
@@ -1,0 +1,1 @@
+export * from './VersionWarningBanner';
