+++ src/shared/components/Banners/index.ts
@@ -1,0 +1,2 @@
+export * from './VersionWarningBanner';
+export * from './Banner';
