+++ src/shared/components/index.ts
@@ -1,7 +1,7 @@
 export * from './Actions';
 export * from './AppletImage';
 export * from './Avatar';
-export * from './Banner';
+export * from './Banners';
 export * from './ButtonWithMenu';
 export * from './Chip';
 export * from './CropPopup';
