+++ src/shared/state/Banners/Banners.reducer.ts
@@ -1,15 +1,15 @@
 import { PayloadAction } from '@reduxjs/toolkit';
 
-import { BannerPayload, BannerSchema } from './Banners.schema';
+import { BannerPayload, BannersSchema } from './Banners.schema';
 
 export const reducers = {
-  addBanner: (state: BannerSchema, { payload }: PayloadAction<BannerPayload>): void => {
+  addBanner: (state: BannersSchema, { payload }: PayloadAction<BannerPayload>): void => {
     state.data.banners.push(payload);
   },
-  removeBanner: (state: BannerSchema, { payload }: PayloadAction<BannerPayload>): void => {
+  removeBanner: (state: BannersSchema, { payload }: PayloadAction<BannerPayload>): void => {
     state.data.banners = state.data.banners.filter(({ key }) => key !== payload.key);
   },
-  removeAllBanners: (state: BannerSchema): void => {
+  removeAllBanners: (state: BannersSchema): void => {
     state.data.banners = [];
   },
 };
